skill-tree/src/main/java/com/RDS/skilltree/utils/UUIDValidationInterceptor.java
package com.RDS.skilltree.utils;

import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import org.springframework.stereotype.Component;
import org.springframework.web.servlet.HandlerInterceptor;

@Component
public class UUIDValidationInterceptor implements HandlerInterceptor {
    @Override
    public boolean preHandle(
            HttpServletRequest request, HttpServletResponse response, Object handler) {
        // Add logic during implementation
        return true;
    }
}
